requirements.txt
aiogram==3.2.0
sqlalchemy==2.0.23
apscheduler==3.10.4
python-dotenv==1.0.0